Add basic support for extracting bonuses from items
EG: Cloak of Protection provides +1 to AC and saving throws.
There's still quite a bit of work to be done here—for example,
Bracers of Archery provide a bonus to attack and damage rolls,
but only with ranged weapons. But this is a good start.

I think the last thing to do before we can go ahead and merge this
code and add items to Wish is to break out multiple weapon types
(EG: Flame Tongue) like we do for Armor, then also the +1, +2, +3
stuff, and that'll be a great start.

Support for charges and providing limited use stuff will probably
come later....